+++ lib/papa/command/larga/deploy.rb
@@ -4,10 +4,8 @@ module Papa
   module Command
     module Larga
       class Deploy < Command::Base
-        RELEASE_OR_HOTFIX_LIFESPAN = '3d'
-        DEFAULT_LIFESPAN = '4h'
-        RELEASE_OR_HOTFIX_PROTECTION = 'off'
-        DEFAULT_PROTECTION = 'on'
+        RELEASE_OR_INTEGRATION_LIFESPAN = '3d'
+        RELEASE_OR_INTEGRATION_PROTECTION = 'off'
 
         def initialize(options)
           @options = options
@@ -21,7 +19,7 @@ module Papa
           @stdout = ''
           IO.popen(command).each do |line|
             @stdout << line
-            puts line
+            Helper::Output.info line
           end
           self
         end
@@ -38,32 +36,45 @@ module Papa
         private
 
         def larga_options
+          action = determine_action
           branch = @options[:branch]
-          lifespan =
-            if branch_is_release_or_hotfix?
-              RELEASE_OR_HOTFIX_LIFESPAN
-            else
-              DEFAULT_LIFESPAN
-            end
-          protection =
-            if branch_is_release_or_hotfix?
-              RELEASE_OR_HOTFIX_PROTECTION
-            else
-              DEFAULT_PROTECTION
-            end
+          lifespan = RELEASE_OR_INTEGRATION_LIFESPAN
+          protection = RELEASE_OR_INTEGRATION_PROTECTION
           hostname = @options[:hostname]
 
           options = []
-          options << '-action build'
+          options << "-action #{action}"
           options << "-branch #{branch}"
           options << "-lifespan #{lifespan}"
           options << "-protection #{protection}"
-          options << "-hostname #{hostname}" if hostname
+          options << "-hostname #{hostname}"
           options
         end
 
-        def branch_is_release_or_hotfix?
-          ['release', 'hotfix'].any? { |s| @options[:branch].include?(s) }
+        def determine_action
+          previous_branches = `larga -action show -branch x | grep https://#{@options[:hostname]} | awk '{print $1}'`.chomp.split("\n")
+          if previous_branches.empty?
+            # If there are no old instances with the same hostname, continue building instance.
+            return 'build'
+          elsif previous_branches.count == 1 && previous_branches.include?(@options[:branch])
+            # If the old instance has the same branch name, skip building and just deploy.
+            return 'deploy'
+          else
+            # If all else fails, destroy old instance(s), and continue building.
+            destroy_old_instances(previous_branches)
+            return 'build'
+          end
+        end
+
+        def destroy_old_instances(previous_branches)
+          Helper::Output.stdout "Destroying old instances..."
+          previous_branches.each do |previous_branch|
+            destroy_command = "larga -action destroy -branch #{previous_branch}"
+            Helper::Output.stdout "Running #{destroy_command.bold}..."
+            IO.popen(destroy_command).each do |line|
+              Helper::Output.info line
+            end
+          end
         end
       end
     end