+++ lib/papa/task/common/deploy.rb
@@ -1,23 +1,14 @@
-require 'papa/helper/output'
+require 'papa/task/base'
 require 'papa/command/larga/type'
 require 'papa/command/larga/deploy'
-require 'papa/runner'
+require 'papa/command/slack/send_message'
 
 module Papa
   module Task
     module Common
-      class Deploy
-        def run
-          @build_branch ||= "#{@build_type}/#{@version}"
-
-          runner = Runner.new(queue)
-
-          if runner.run
-            success_message
-          else
-            failure_message
-            exit 1
-          end
+      class Deploy < Base
+        def initialize
+          check_if_build_branch_exists
         end
 
         private
@@ -25,13 +16,15 @@ module Papa
         def queue
           [
             Command::Larga::Type.new,
-            Command::Larga::Deploy.new(deploy_options)
+            Command::Slack::SendMessage.new(@build_type, @hostname, 'started'),
+            Command::Larga::Deploy.new(deploy_options),
+            Command::Slack::SendMessage.new(@build_type, @hostname, 'done')
           ]
         end
 
         def deploy_options
           {
-            branch: @build_branch,
+            branch: build_branch,
             hostname: @hostname
           }
         end
@@ -39,7 +32,7 @@ module Papa
         def success_message
           Helper::Output.success 'Successfully deployed larga instance.'
           info = ''
-          info << "  Branch: #{@build_branch}\n"
+          info << "  Branch: #{build_branch}\n"
           info << "  URL: https://#{@hostname}.indinerocorp.com\n"
           Helper::Output.success_info info
         end
@@ -47,7 +40,7 @@ module Papa
         def failure_message
           Helper::Output.failure 'There was a problem deploying larga instance.'
           info = ''
-          info << "  Branch: #{@build_branch}\n"
+          info << "  Branch: #{build_branch}\n"
           Helper::Output.failure_info info
         end
       end